Revitalised the old atom-decomposition code + plotting system;
Now outputs to its own file like a sensible routine.
And briefly fiddled with mode_decomposition.gpt to get it to run OK and
remove some of the more complex options, while still doing nice things
such as autogenerating the coloured header, and putting the frequency of
the mode side-plotted at the top.
Currently it still has a fair few hard-codes for MA.PbX3 like
structures, in terms of colouring + number of modes plotted.